refactor: Line chart and Bar chart
