+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetBarChart/components/BarChartEndLines.tsx
@@ -1,0 +1,46 @@
+import { type ComputedBarDatum } from '@nivo/bar';
+import { isDefined } from 'twenty-shared/utils';
+import { type BarChartDataItem } from '../types/BarChartDataItem';
+import { type BarChartEnrichedKey } from '../types/BarChartEnrichedKey';
+import { calculateBarEndLineCoordinates } from '../utils/calculateBarEndLineCoordinates';
+
+type BarChartEndLinesProps = {
+  bars: readonly ComputedBarDatum<BarChartDataItem>[];
+  enrichedKeys: BarChartEnrichedKey[];
+  layout: 'vertical' | 'horizontal';
+};
+
+export const BarChartEndLines = ({
+  bars,
+  enrichedKeys,
+  layout,
+}: BarChartEndLinesProps) => {
+  if (!bars || bars.length === 0) {
+    return null;
+  }
+
+  return (
+    <g>
+      {bars.map((bar: ComputedBarDatum<BarChartDataItem>, index: number) => {
+        const enrichedKey = enrichedKeys.find((k) => k.key === bar.data.id);
+        if (!isDefined(enrichedKey)) {
+          return null;
+        }
+        const lineColor = enrichedKey.colorScheme.solid;
+        const { x1, y1, x2, y2 } = calculateBarEndLineCoordinates(bar, layout);
+
+        return (
+          <line
+            key={`${bar.data.id}-${bar.data.indexValue}-endline-${index}`}
+            x1={x1}
+            y1={y1}
+            x2={x2}
+            y2={y2}
+            stroke={lineColor}
+            strokeWidth={1}
+          />
+        );
+      })}
+    </g>
+  );
+};
