+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetBarChart/hooks/useBarChartHandlers.ts
@@ -1,0 +1,35 @@
+import { useState } from 'react';
+import { type ComputedDatum, type BarDatum } from '@nivo/bar';
+import { isDefined } from 'twenty-shared/utils';
+import { type BarChartDataItem } from '../types/BarChartDataItem';
+
+type UseBarChartHandlersProps = {
+  data: BarChartDataItem[];
+  indexBy: string;
+};
+
+export const useBarChartHandlers = ({
+  data,
+  indexBy,
+}: UseBarChartHandlersProps) => {
+  const [hoveredBar, setHoveredBar] = useState<{
+    key: string;
+    indexValue: string | number;
+  } | null>(null);
+
+  const handleBarClick = (datum: ComputedDatum<BarDatum>) => {
+    const dataItem = data.find((d) => d[indexBy] === datum.indexValue);
+    if (isDefined(dataItem?.to)) {
+      window.location.href = dataItem.to;
+    }
+  };
+
+  const hasClickableItems = data.some((item) => isDefined(item.to));
+
+  return {
+    hoveredBar,
+    setHoveredBar,
+    handleBarClick,
+    hasClickableItems,
+  };
+};
