+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetBarChart/types/BarChartConfig.ts
@@ -1,0 +1,8 @@
+import { type GraphColorScheme } from '../../../types/GraphColorScheme';
+
+export type BarChartConfig = {
+  key: string;
+  indexValue: string | number;
+  gradientId: string;
+  colorScheme: GraphColorScheme;
+};
