+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetBarChart/types/BarChartDataItem.ts
@@ -1,0 +1,5 @@
+import { type BarDatum } from '@nivo/bar';
+
+export type BarChartDataItem = BarDatum & {
+  to?: string;
+};
