+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetBarChart/types/BarChartEnrichedKey.ts
@@ -1,0 +1,7 @@
+import { type GraphColorScheme } from '../../../types/GraphColorScheme';
+
+export type BarChartEnrichedKey = {
+  key: string;
+  colorScheme: GraphColorScheme;
+  label: string;
+};
