+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetBarChart/types/BarChartSeries.ts
@@ -1,0 +1,7 @@
+import { type GraphColor } from '../../../types/GraphColor';
+
+export type BarChartSeries = {
+  key: string;
+  label?: string;
+  color?: GraphColor;
+};
