+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetBarChart/utils/getBarColor.ts
@@ -1,0 +1,17 @@
+import { type ComputedDatum, type BarDatum } from '@nivo/bar';
+import { type ThemeType } from 'twenty-ui/theme';
+import { type BarChartConfig } from '../types/BarChartConfig';
+
+export const getBarColor = (
+  datum: ComputedDatum<BarDatum>,
+  barConfigs: BarChartConfig[],
+  theme: ThemeType,
+) => {
+  const bar = barConfigs.find(
+    (b) => b.key === datum.id && b.indexValue === datum.indexValue,
+  );
+  if (!bar) {
+    return theme.border.color.light;
+  }
+  return `url(#${bar.gradientId})`;
+};
