+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetLineChart/types/LineChartDataPoint.ts
@@ -1,0 +1,5 @@
+export type LineChartDataPoint = {
+  x: number | string | Date;
+  y: number | null;
+  to?: string;
+};
