+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetLineChart/types/LineChartEnrichedSeries.ts
@@ -1,0 +1,9 @@
+import { type GraphColorScheme } from '../../../types/GraphColorScheme';
+import { type LineChartSeries } from './LineChartSeries';
+
+export type LineChartEnrichedSeries = LineChartSeries & {
+  colorScheme: GraphColorScheme;
+  gradientId: string;
+  shouldEnableArea: boolean;
+  label: string;
+};
