+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetLineChart/types/LineChartSeries.ts
@@ -1,0 +1,10 @@
+import { type GraphColor } from '../../../types/GraphColor';
+import { type LineChartDataPoint } from './LineChartDataPoint';
+
+export type LineChartSeries = {
+  id: string;
+  label?: string;
+  color?: GraphColor;
+  data: LineChartDataPoint[];
+  enableArea?: boolean;
+};
