+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetLineChart/utils/getAxisBottomConfig.ts
@@ -1,0 +1,8 @@
+export const getAxisBottomConfig = (xAxisLabel?: string) => ({
+  tickSize: 0,
+  tickPadding: 5,
+  tickRotation: 0,
+  legend: xAxisLabel,
+  legendPosition: 'middle' as const,
+  legendOffset: 40,
+});
