+++ packages/twenty-front/src/modules/page-layout/widgets/graph/components/GraphWidgetLineChart/utils/getAxisLeftConfig.ts
@@ -1,0 +1,17 @@
+import {
+  formatGraphValue,
+  type GraphValueFormatOptions,
+} from '../../../utils/graphFormatters';
+
+export const getAxisLeftConfig = (
+  yAxisLabel?: string,
+  formatOptions?: GraphValueFormatOptions,
+) => ({
+  tickSize: 0,
+  tickPadding: 5,
+  tickRotation: 0,
+  legend: yAxisLabel,
+  legendPosition: 'middle' as const,
+  legendOffset: -50,
+  format: (value: number) => formatGraphValue(value, formatOptions || {}),
+});
